Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    float a;
-    scanf("%f",&a);
-    printf("You entered: %.4f\n",a);
+    double a;
+    scanf("%lf",&a);
+    printf("You entered: %.4lf\n",a);
     return 0;
 }
